scripts/figures/approx_reference_front/requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
jmetalpy==1.9.0
